requirements.txt
asgiref==3.9.1
Django==5.2.4
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0

# Logging and monitoring
python-json-logger==3.3.0
django-prometheus==2.3.1
prometheus-client==0.20.0
gunicorn==23.0.0

# Caching
django-cache-memoize==0.1.10

# Audit logging
django-auditlog==2.3.0

# Logstash integration
python-logstash-async==4.0.2

# Elasticsearch client
elasticsearch==9.1.0
